[stabilizer] Split stabilizer and logging/gui implementation
